backend/requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy
python-jose[cryptography] 
passlib[bcrypt] 
sqlalchemy
pydantic[email]